Fix invalid pyproject.toml file and move to setup.cfg (#201)
The `[project]` section of a `pyproject.toml` file *has* to include both
a `name` and a `version`, that we were missing.

However, rather than adding more stuff to `pyproject.toml`, I've moved
the license info to the setup file and pointed `pyproject.toml` to use
`setuptools` for building (the `pyproject.toml` file is really only
needed to integrated with `black`, that doesn't support setup files...)

While I was at it I decided to migrate to a `setup.cfg` file rather than
`setup.py` since its declaritive config is usally preferred these days
(and it's a bit easier to load from version/license files than with a
`setup.py` file).